SITES-16873: [TBYB] Create an AEM Headless Trial signup site built with Edge Delivery Services
- update form block
- fix styles
- bring in old content
