+++ blocks/form/form.js
@@ -1,77 +1,61 @@
 function constructPayload(form) {
-  const payload = {};
-  [...form.elements].forEach((fe) => {
-    if (fe.id) {
-      payload[fe.id] = fe.value;
+  return [...form.elements].reduce((payload, formElement) => {
+    if (formElement.id) {
+      payload[formElement.id] = formElement.value;
     }
-  });
-  return payload;
+    return payload;
+  }, {});
 }
 
-/**
- * Once the payload is constructed and the form entry is validated,
- * we submit the form. It creates an entry in the spreadsheet to store the form
- * data.
- */
-async function submitForm(form) {
+async function handleSubmit(form) {
   const payload = constructPayload(form);
-
+  payload.timestamp = new Date().toJSON();
   const resp = await fetch(form.dataset.action, {
     method: 'POST',
     cache: 'no-cache',
-    headers: {
-      'Content-Type': 'application/json',
-    },
+    headers: { 'Content-Type': 'application/json' },
     body: JSON.stringify({ data: payload }),
   });
   await resp.text();
   return payload;
 }
 
-/**
- * We check the validity of the email entry. We only accept company email addresses
- * as of now.
- */
 function checkValidity(form) {
   const payload = constructPayload(form);
   const { email } = payload;
-  const regex = /^[^@]+@(?!(yahoo|hotmail|gmail|icloud|outlook))[^@]+\.[a-z]{2,}$/;
+  const regex = /^[_A-Za-z0-9-+]+(\.[_A-Za-z0-9-]+)*@(?!(yahoo|hotmail|gmail|icloud|outlook))[^@]+\.[a-z]{2,}$/;
   return String(email).match(regex);
 }
 
-/**
- * We create a submit button for the form.
- * Once we click on submit, if the email entry is valid, we disable
- * the submit button and redirect the user to the thank you page.
- */
 function createButton(field) {
   const button = document.createElement('button');
   button.textContent = field.Label;
-  button.classList.add('button');
-  button.classList.add('primary');
+  button.classList.add('con-button', 'button-l', 'fill');
   if (field.Type === 'submit') {
     button.addEventListener('click', async (event) => {
       const form = button.closest('form');
+      const block = button.closest('.form');
       event.preventDefault();
       if (checkValidity(form) !== null) {
         button.setAttribute('disabled', '');
-        await submitForm(form);
-        const formContainer = form.closest('.form-container');
-        formContainer.firstElementChild.classList.add('hide');
-        form.classList.add('hide');
+        await handleSubmit(form);
+        block.classList.add('submitted');
+        const formParent = form.parentElement;
         const paragraph = document.createElement('p');
-        paragraph.classList.add('trial-sign-up-message');
+        paragraph.classList.add('form-submit-message');
         const text = document.createTextNode(field.Extra);
         paragraph.appendChild(text);
-        formContainer.appendChild(paragraph);
+        formParent.appendChild(paragraph);
       } else {
-        const emailElement = document.getElementById('email');
-        if (emailElement.closest('div').childNodes.length < 2) {
+        const emailElement = form.querySelector('#email');
+        if (emailElement?.closest('.field-wrapper').childNodes.length < 2) {
           const paragraph = document.createElement('p');
-          const text = document.createTextNode(field.Extra);
+          const text = document.createTextNode(field.Error);
+          const label = emailElement.closest('.form-label');
+          label.classList.add('alert');
           paragraph.appendChild(text);
           emailElement.classList.add('highlight');
-          emailElement.closest('div').appendChild(paragraph);
+          emailElement.closest('.field-wrapper').appendChild(paragraph);
         }
       }
     });
@@ -79,33 +63,34 @@ function createButton(field) {
   return button;
 }
 
-/**
- * We create an input field and pick up the type, id, and placeholder text
- * from the helix-default spreadsheet on sharepoint.
- */
 function createInput(field) {
   const input = document.createElement('input');
   input.type = field.Type;
   input.id = field.Field;
-  input.setAttribute('placeholder', field.Placeholder);
-  if (field.Mandatory === 'x') {
-    input.setAttribute('required', 'required');
+  input.name = field.Field;
+
+  const search = new URLSearchParams(window.location.search);
+  const searchValue = search.get(field.Field);
+  if (searchValue) {
+    input.value = searchValue;
+  } else if (field.Value) {
+    input.value = field.Value;
   }
+
+  if (field.Type !== 'hidden') {
+    input.setAttribute('placeholder', field.Placeholder);
+    if (field.Mandatory === 'x') {
+      input.setAttribute('required', 'required');
+    }
+    const label = document.createElement('label');
+    label.classList.add('form-label');
+    label.appendChild(input);
+    return label;
+  }
+
   return input;
 }
 
-function fill(form) {
-  const { action } = form.dataset;
-  if (action === '/tools/bot/register-form') {
-    const loc = new URL(window.location.href);
-    form.querySelector('#owner').value = loc.searchParams.get('owner') || '';
-    form.querySelector('#installationId').value = loc.searchParams.get('id') || '';
-  }
-}
-
-/**
- * We create the form and add the submit button and input text field to it.
- */
 async function createForm(formURL) {
   const { pathname } = new URL(formURL);
   const resp = await fetch(pathname);
@@ -114,11 +99,8 @@ async function createForm(formURL) {
   // eslint-disable-next-line prefer-destructuring
   form.dataset.action = pathname.split('.json')[0];
   json.data.forEach((field) => {
-    field.Type = field.Type || 'text';
     const fieldWrapper = document.createElement('div');
-    const style = field.Style ? ` form-${field.Style}` : '';
-    const fieldId = `form-${field.Type}-wrapper${style}`;
-    fieldWrapper.className = fieldId;
+    fieldWrapper.className = `form-${field.Type}-wrapper`;
     fieldWrapper.classList.add('field-wrapper');
     switch (field.Type) {
       case 'submit':
@@ -130,13 +112,14 @@ async function createForm(formURL) {
     form.append(fieldWrapper);
   });
 
-  fill(form);
   return form;
 }
 
-export default async function decorate(block) {
-  const form = block.querySelector('a[href$=".json"]');
-  if (form) {
-    form.replaceWith(await createForm(form.href));
+const init = async (el) => {
+  const anchor = el.querySelector('p > a[href$=".json"]');
+  if (anchor) {
+    anchor.parentElement.replaceWith(await createForm(anchor.href));
   }
-}
+};
+
+export default init;
