--- blocks/trial-signup-form/trial-signup-form.css
@@ -1,12 +1,0 @@
-.trial-signup-form {
-  display: none;
-}
-
-.trial-signup-form.visible {
-  display: block;
-}
-
-.trial-signup-form iframe {
-  width: 100%;
-  border: none;
-}
